ZKLedger compatibility work
This commit contains changes needed for integrating ZKSigma into
ZKLedger:

* The ZKCurve is no longer in here, but you need to pass the curve
parameters into the proof functions. ZKCurve will be defined in
ZKLedger. There is a TestCurve defined that is used in the test cases.

* ConsistencyProof struct contained two private variables; that resulted
in failures when ZKLedger serialized/deserialized these after
transmitting them over the wire. Resulting in failures in the local
experiments.